// fmea-web-client/src/libs/api-client/model/fMStructureDto2.ts
/**
 * Fmea API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FMStructureDto2 { 
    code: string;
    longName: string;
    shortName: string;
    category: string;
    seq: number;
    level: number;
    decomposition: Array<string>;
    functions: Array<string>;
}